Add react vite plugin if not found
