+++ code/frameworks/react-vite/package.json
@@ -55,7 +55,7 @@
     "@rollup/pluginutils": "^4.2.0",
     "@storybook/builder-vite": "7.0.0-alpha.33",
     "@storybook/react": "7.0.0-alpha.33",
-    "@vitejs/plugin-react": "^2.0.1",
+    "@vitejs/plugin-react": "^2.0.0",
     "ast-types": "^0.14.2",
     "magic-string": "^0.26.1",
     "react-docgen": "^6.0.0-alpha.3",
