+++ code/frameworks/react-vite/src/preset.ts
@@ -1,7 +1,6 @@
 /* eslint-disable global-require */
-import path from 'path';
-import fs from 'fs';
 import type { StorybookConfig, TypescriptOptions } from '@storybook/builder-vite';
+import { hasPlugin, readPackageJson } from './utils';
 
 export const addons: StorybookConfig['addons'] = ['@storybook/react'];
 
@@ -9,25 +8,21 @@ export const core: StorybookConfig['core'] = {
   builder: '@storybook/builder-vite',
 };
 
-export function readPackageJson(): Record<string, any> | false {
-  const packageJsonPath = path.resolve('package.json');
-  if (!fs.existsSync(packageJsonPath)) {
-    return false;
-  }
-
-  const jsonContent = fs.readFileSync(packageJsonPath, 'utf8');
-  return JSON.parse(jsonContent);
-}
-
 export const viteFinal: StorybookConfig['viteFinal'] = async (config, { presets }) => {
   const { plugins = [] } = config;
 
+  // Add react plugin if not present
+  if (!hasPlugin(plugins, 'vite:react-babel')) {
+    const { default: react } = await import('@vitejs/plugin-react');
+    plugins.push(react());
+  }
+
+  // Add docgen plugin
   const { reactDocgen, reactDocgenTypescriptOptions } = await presets.apply<any>(
     'typescript',
     {} as TypescriptOptions
   );
   let typescriptPresent;
-
   try {
     const pkgJson = readPackageJson();
     typescriptPresent =
@@ -35,7 +30,6 @@ export const viteFinal: StorybookConfig['viteFinal'] = async (config, { presets 
   } catch (e) {
     typescriptPresent = false;
   }
-
   if (reactDocgen === 'react-docgen-typescript' && typescriptPresent) {
     plugins.push(
       require('@joshwooding/vite-plugin-react-docgen-typescript')(reactDocgenTypescriptOptions)
